Add includes Remove ctors from structs
