+++ modules/backend/common/enumerations/allergen.hpp
@@ -8,7 +8,8 @@ enum class Allergen : uint8_t
     Allergen = 0,
     AnotherAllergen = 1,
     SomeAllergen = 2,
-    DifferentAllergen = 3
+    DifferentAllergen = 3,
+    Unknown = 255
 };
 
 std::string toString(const Allergen& shift)
