+++ modules/backend/common/enumerations/specialization.hpp
@@ -6,6 +6,7 @@
 enum class Specialization : uint8_t
 {
     Dentist = 0,
+    Unknown = 255
 };
 
 std::string toString(const Specialization& specialization)
