+++ modules/backend/common/enumerations/treatment_state.hpp
@@ -6,7 +6,8 @@
 enum class TreatmentState : uint8_t
 {
     Planned = 0,
-    Done = 1
+    Done = 1,
+    Unknown = 255
 };
 
 std::string toString(const TreatmentState& treatment_state)
