+++ modules/backend/common/personal_data/address/address.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <string>
+
+struct Address
+{
+    std::string city_;
+    std::string country_;
+    std::string province_;
+    std::string street_;
+    std::string zip_code_;
+};
