+++ modules/backend/common/personal_data/name/name.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <string>
+
+struct Name
+{
+    std::string first_name_;
+    std::string second_name_;
+    std::string last_name_;
+};
