+++ modules/backend/common/personal_data/phone_number/phone_number.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <string>
+
 struct PhoneNumber
 {
     std::string phone_code_; // +1, +48 (+ == 00)
