+++ modules/backend/visits/visit/visit.hpp
@@ -1,19 +1,23 @@
 #pragma once
 
+#include "modules/backend/patients/patient/i_patient.hpp"
+#include <vector>
+
 class Visit
 {
-    iVisitPatient patient_;
-    std::vector<Treatment> actual_treatment_;
+
+    // std::vector<Treatment> actual_treatment_;
     // on gui, doctor should select treatment/s for exact visit
   public:
-    Visit(const Patient& patient) : patient_{patient}
-    {
-    }
-
+  // used to set informations required for visit from patient
+  void SetPatientInfo(const iPatient& patient) 
+  {
+    // patient.GetPatientInfo();
+  }
     void FinishVisit()
     {
-        const TreatmentState treatment_state = TreatmentState::Done;
-        patient_.UpdateTreatment(treatment_state, actual_treatment_);
+        // const TreatmentState treatment_state = TreatmentState::Done;
+        // patient_.UpdateTreatment(treatment_state, actual_treatment_);
     }
 };
